fix: Input 경로 수정
