+++ frontend/index.html
@@ -33,6 +33,7 @@
 	<link rel="stylesheet" href="./src/components/FriendGameRecords/FriendOneGameRecord.css">
 	<link rel="stylesheet" href="./src/components/FriendGameRecords/FriendOutcome.css">
 	<link rel="stylesheet" href="./src/components/Modal/modal.css">
+	<link rel="stylesheet" href="./src/components/Input/Input.css">
 </head>
 
 <body>
