+++ frontend/src/components/Input/Input.css
@@ -1,0 +1,17 @@
+.input-default {
+	border: 1px solid #ccc;
+}
+
+.input-background {
+	border: 1px solid #000;
+	background-color: #fff;
+}
+
+.small {
+}
+
+.medium {
+}
+
+.large {
+}
