+++ frontend/src/components/JoinModal/JoinModal.js
@@ -1,7 +1,7 @@
 import { Modal } from "../Modal/index.js";
 import { SelectCharactor } from "../SelectCharactor/SelectCharactor.js";
 import { Button } from "../Button.js";
-import { Input } from "../Input.js";
+import { Input } from "../../components/Input/Input.js";
 
 export class JoinModal extends Modal {
     constructor(onClose = () => { }) {
